refactor: create a component for certificates carrousel
