+++ new-portifolio/app/Types/Certificates.ts
@@ -1,0 +1,7 @@
+import { StaticImageData } from "next/image";
+
+export interface Certificates {
+  image: string | StaticImageData;
+  title: string;
+  description: string
+}
